Reduce spec tests boilerplate
