--- packages/spec-test-runner/test/spec/altair/epoch_processing/effective_balance_updates/mainnet.test.ts
@@ -1,4 +1,0 @@
-import {runEffectiveBalanceUpdates} from "./effective_balance_updates";
-import {PresetName} from "@chainsafe/lodestar-params";
-
-runEffectiveBalanceUpdates(PresetName.mainnet);
